added base classes and new handlers
abc created for users and channels
there is now an event handler which can pass events or messages to the right locations.
the message is passed around which has all the goodies needed to be sent by the client.
the message has meta data so that its history can be tracked.
updated example.py to use the new message class